Added various options
# NOTE: Version still requires further refactoring. A lot of unnecessary & messy code is still left to be fixed.

Various options were added, allowing more control by the end-user for a given table of contents.

With this many options, it might be a good idea to have settings within the 'Community plugins' panel/UI.
This to allow the default behaviour to be changed, should a user want to use the same non-default settings for every ATOC.
Would have to look into how that works, but it sounds more than doable.

# AllowStyleHTML (bool )
| Allows a heading to be styled by HTML only when includeLinks is set to false.
| Default value: false
| ( Credit to jayv in the initial issue, for the HTML-stripping regex :trophy: )

# includeMarkdownLinks (bool)
| Allows Markdown links in heading only when includeLinks is set to false.
| Default value: true

# showMarkdownLinksInLinks (bool)
| Keeps the link destination from markdown links visible in headings while includeLinks is set to true.
| Default value: false